Complete divergences
however the implementation of kl-divergencedoes not match the result of implementation
from dit, yet the result yields the same as the example shows from
kl-divergence wiki-pedia, which reuslt in symmetric kl-divergence and
jensen-shannon divergence implementation also has different result with
dit. We will assume that the current version is not errornous.